remoteproc: qcom: mdt_loader: Fix rproc_da_to_va
Refactoring MDT loader broke rproc_da_to_va() for Qualcomm images.  This
causes an incorrect computation of offset and returns an incorrect
virtual address for the relocatable segment(s).

Fix this by adding an "out" parameter to qcom_mdt_loader() and setting
the correct segment base address in that parameter.

Fixes: 7f0dd07a9b29 ("remoteproc: qcom: mdt_loader: Refactor MDT loader")
CC: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Sarangdhar Joshi <spjoshi@codeaurora.org>